update liburing-rs to support Alpine
Support for Alpine Linux proved challening on multiple different levels.
Because Alpine doesn't seem to support dynamically opening shared
libraries during runtime, bindgen is not feasible as a crate dependency,
which only leaves the CLI version.

We update the build.rs to accommodate a new workflow where bindgen from
the command line is used instead, spawned as a new child process. This
is sufficient on Alpine where it's trivial to install it as:

  apk add rust-bindgen

We add the new feature `bindgen-cli` to the main crate along with the
dependent testing crate.